vmake to enhance make.exe called by Eclipse C/C++
This console app is placed in the same folder as make.exe and eliminates this issue by waiting for the pre-build phase to complete before invoking the main build list.  You're welcome to use and modify as needed.  Set the Eclipse build to use vmake.exe rather than make.exe.